+++ settings.php
@@ -43,12 +43,11 @@ if ($ADMIN->fulltree) {
         ['username' => 1],
         $choices
     ));
-
     $settings->add(new admin_setting_configcheckbox(
         'report_advancedgrading/enable_javascriptlayout',
         get_string('enable_javascriptlayout', 'report_advancedgrading'),
         get_string('enable_javascriptlayout_desc', 'report_advancedgrading'),
-        '0'
+        '1'
     ));
 
 }