+++ changelog.md
@@ -1,3 +1,7 @@
-### Version 0.1
+### Version 0.2 Oct 2022
+Removed spurious % in time graded output.
+Thanks to Vlad Kidanov for reporting the alphabet generation would stop at 26 and contributing
+code to make it dependent on column count.
 
-
+### Version 0.1 Feb 2022
+First release